requirements/comm.txt
django~=3.0.0
pillow
djangorestframework
django-cors-headers
psycopg2
djangorestframework-jwt
celery[redis]